# use MemoryMarshal.GetReference to implement FromFixedSpan