Note:

- makes the `Span<byte>` and `ReadOnlySpan<byte>` paths consistent
- avoids `fixed` stack gynmastics overheads (minor but non-zero)
- works correctly for zero-length spans (i.e. doesn't give `IndexOutOfRangeException`)