@badrishc 2024-04-02
Need this in Tsavorite ...